debug a little -by hongfei 1:24pm jun 26
